# modules/RestBuilder/RestBuilder-impl/rest-openapi.yaml
components:
  schemas:
    Category:
      description: Category
      properties:
        name:
          description: The category name.
          type: string
        id:
          description: The category ID.
          type: integer
        tax:
          description: The tax category.
          format: double
          type: number
      type: object
...
info:
  description: ""
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
  title: "Sale Taxe API"
  version: v1.0
openapi: 3.0.1
paths:
...
  "/category/{categoryName}":
    get:
      operationId: getCategory
      parameters:
        - in: path
          name: categoryName
          required: true
          schema:
            type: string
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
            application/xml:
              schema:
                $ref: "#/components/schemas/Category"
          description: ""
      tags: ["Category"]
  "/category/{categoryId}":
    get:
      parameters:
        - in: path
          name: categoryId
          required: true
          schema:
            type: integer
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
            application/xml:
              schema:
                $ref: "#/components/schemas/Category"
          description: ""
      tags: [ "Category" ]
  "/category/post":
    post:
      operationId: postCategoryPost
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Category"
          application/xml:
            schema:
              $ref: "#/components/schemas/Category"
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
            application/xml:
              schema:
                $ref: "#/components/schemas/Category"
          description: ""
      tags: ["Category"]
  "/category/delete/{categoryId}":
    delete:
      operationId: deleteCategoryDeleteCategory
      parameters:
        - in: path
          name: categoryId
          required: true
          schema:
            type: integer
      responses:
        204:
          content:
            application/json: {}
            application/xml: {}
          description: ""
      tags: ["Category"]
...